+++ inst/app/init.R
@@ -69,84 +69,125 @@ r_ssuid <- if (getOption("radiant.local")) {
 ## (re)start the session and push the id into the url
 session$sendCustomMessage("session_start", r_ssuid)
 
-## load for previous state if available but look in global memory first
-# if (exists("r_data")) {
-#   r_data <- do.call(reactiveValues, r_data)
-#   r_state <- if (exists("r_state")) r_state else list()
-#   suppressWarnings(rm(r_data, r_state, envir = .GlobalEnv))
-# } else if (!is.null(r_sessions[[r_ssuid]]$r_data)) {
-#   r_data <- do.call(reactiveValues, r_sessions[[r_ssuid]]$r_data)
-#   r_state <- r_sessions[[r_ssuid]]$r_state
-# } else if (file.exists(paste0("~/radiant.sessions/r_", r_ssuid, ".rds"))) {
-#   ## read from file if not in global
-#   fn <- paste0(normalizePath("~/radiant.sessions"), "/r_", r_ssuid, ".rds")
+## identify the shiny environment
+r_environment <- environment()
+## create a child environment to use for Report > Rmd and Report > R
+# r_data <- environment()
 
-#   rs <- try(readRDS(fn), silent = TRUE)
-#   if (is(rs, "try-error")) {
-#     r_data <- init_data()
-#     r_state <- list()
-#   } else {
-#     if (length(rs$r_data) == 0) {
-#       r_data <- init_data()
-#     } else {
-#       r_data <- do.call(reactiveValues, rs$r_data)
-#     }
+listToEnv <- function(x, envir = new.env()) {
+  if (is.list(x)) {
 
-#     if (length(rs$r_state) == 0) {
-#       r_state <- list()
-#     } else {
-#       r_state <- rs$r_state
-#     }
-#   }
-
-#   unlink(fn, force = TRUE)
-#   rm(rs)
-# } else if (isTRUE(getOption("radiant.local")) && file.exists(paste0("~/radiant.sessions/r_", mrsf, ".rds"))) {
-
-#   ## restore from local folder but assign new ssuid
-#   fn <- paste0(normalizePath("~/radiant.sessions"), "/r_", mrsf, ".rds")
-#   rs <- try(readRDS(fn), silent = TRUE)
-#   if (is(rs, "try-error")) {
-#     r_data <- init_data()
-#     r_state <- list()
-#   } else {
-#     r_data <- if (length(rs$r_data) == 0) init_data() else do.call(reactiveValues, rs$r_data)
-#     r_state <- if (length(rs$r_state) == 0) list() else rs$r_state
-#   }
-
-#   ## don't navigate to same tab in case the app locks again
-#   r_state$nav_radiant <- NULL
-
-#   unlink(fn, force = TRUE)
-#   rm(rs)
-# } else {
-#   r_data <- init_data()
-#   r_state <- list()
-# }
-
-knitr_environment <- new.env()
-r_data <- knitr_environment
-
-  df_names <- getOption("radiant.init.data", default = c("diamonds", "titanic"))
-  for (dn in df_names) {
-    if (file.exists(dn)) {
-      df <- load(dn) %>% get()
-      dn <- basename(dn) %>%
-        {gsub(paste0(".", tools::file_ext(.)), "", ., fixed = TRUE)}
-    } else {
-      df <- data(list = dn, package = "radiant.data", envir = environment()) %>% get()
-    }
-    r_data[[dn]] <- df
-    r_data[[paste0(dn, "_descr")]] <- attr(df, "description")
-    makeReactiveBinding(dn, env = r_data)
   }
-  r_data$datasetlist <- basename(df_names)
-  r_data$url <- NULL
 
-# print(r_data)
-# init_data()
-print(r_data)
-r_state <- list()
+}
+
+## load for previous state if available but look in global memory first
+if (exists("r_data", envir = .GlobalEnv)) {
+  # print("here global")
+  # probably need to check that all DF in datasetlist are reactive
+  # r_data <- do.call(reactiveValues, r_data)
+  # r_data <- list2env(r_data, envir = new.env())
+  r_data <- if (is.list(r_data)) list2env(r_data, envir = new.env()) else r_data
+  r_state <- if (exists("r_state")) r_state else list()
+  suppressWarnings(rm(r_data, r_state, envir = .GlobalEnv))
+} else if (!is.null(r_sessions[[r_ssuid]]$r_data)) {
+  # print("here r_session")
+  # r_data <- do.call(reactiveValues, r_sessions[[r_ssuid]]$r_data)
+  r_data <- r_sessions[[r_ssuid]]$r_data %>%
+    {if (is.list(.)) list2env(., envir = new.env()) else .}
+  r_state <- r_sessions[[r_ssuid]]$r_state
+} else if (file.exists(paste0("~/radiant.sessions/r_", r_ssuid, ".rds"))) {
+  ## read from file if not in global
+  # print("here radiant.sessions1")
+  fn <- paste0(normalizePath("~/radiant.sessions"), "/r_", r_ssuid, ".rds")
+  rs <- try(readRDS(fn), silent = TRUE)
+  if (is(rs, "try-error")) {
+    r_data <- new.env()
+    init_data(env = r_data)
+    r_state <- list()
+  } else {
+    if (length(rs$r_data) == 0) {
+      r_data <- new.env()
+      init_data(env = r_data)
+    } else {
+      # r_data <- do.call(reactiveValues, rs$r_data)
+      # r_data <- list2env(rs$r_data, envir = new.env())
+      r_data <- rs$r_data  %>%
+        {if (is.list(.)) list2env(., envir = new.env()) else .}
+    }
+
+    if (length(rs$r_state) == 0) {
+      r_state <- list()
+    } else {
+      r_state <- rs$r_state
+    }
+  }
+
+  unlink(fn, force = TRUE)
+  rm(rs)
+} else if (isTRUE(getOption("radiant.local")) && file.exists(paste0("~/radiant.sessions/r_", mrsf, ".rds"))) {
+  # print("here radiant.sessions2")
+  ## restore from local folder but assign new ssuid
+  fn <- paste0(normalizePath("~/radiant.sessions"), "/r_", mrsf, ".rds")
+  rs <- try(readRDS(fn), silent = TRUE)
+  if (is(rs, "try-error")) {
+    r_data <- new.env()
+    init_data(env = r_data)
+    # r_data <- init_data()
+    r_state <- list()
+  } else {
+    # r_data <- if (length(rs$r_data) == 0) init_data() else do.call(reactiveValues, rs$r_data)
+    if (length(rs$r_data) == 0) {
+      r_data <- new.env()
+      init_data(env = r_data)
+    } else {
+      # r_data <- list2env(rs$r_data, envir = new.env())
+      r_data <- rs$r_data %>%
+        {if (is.list(.)) list2env(., envir = new.env()) else .}
+    }
+    r_state <- if (length(rs$r_state) == 0) list() else rs$r_state
+  }
+
+  ## don't navigate to same tab in case the app locks again
+  r_state$nav_radiant <- NULL
+
+  unlink(fn, force = TRUE)
+  rm(rs)
+} else {
+  r_data <- new.env()
+  init_data(env = r_data)
+  r_state <- list()
+}
+
+# if (!exists("datasetlist", envir = r_data)) {
+  # print("making datasetlist")
+  # toReactive("datasetlist")
+isolate({
+  makeReactiveBinding("datasetlist", env = r_data)
+  for (ds in r_data$datasetlist) {
+    makeReactiveBinding(ds, env = r_data)
+  }
+  makeReactiveBinding("dtree_list", env = r_data)
+  for (dt in r_data$dtree_list) {
+    makeReactiveBinding(dt, env = r_data)
+  }
+})
+
+  # df_names <- getOption("radiant.init.data", default = c("diamonds", "titanic"))
+  # for (dn in df_names) {
+  #   if (file.exists(dn)) {
+  #     df <- load(dn) %>% get()
+  #     dn <- basename(dn) %>%
+  #       {gsub(paste0(".", tools::file_ext(.)), "", ., fixed = TRUE)}
+  #   } else {
+  #     df <- data(list = dn, package = "radiant.data", envir = environment()) %>% get()
+  #   }
+  #   r_data[[dn]] <- df
+  #   r_data[[paste0(dn, "_descr")]] <- attr(df, "description")
+  #   makeReactiveBinding(dn, env = r_data)
+  # }
+  # r_data$datasetlist <- basename(df_names)
+  # r_data$url <- NULL
 
 ## legacy, to deal with state files created before
 ## Report > Rmd and Report > R name change
@@ -181,13 +222,6 @@ if (!is.null(r_state$rcode_edit) && is.null(r_state$r_edit)) {
   r_state$r_edit <- r_state$rcode_edit
   r_state$rcode_edit <- NULL
 }
-
-## identify the shiny environment
-r_environment <- environment()
-## create a child environment to use for Report > Rmd and Report > R
-# knitr_environment <- environment()
-# print(r_environment)
-# print(parent.env(knitr_environment))
 
 ## parse the url and use updateTabsetPanel to navigate to the desired tab
 ## currently only works with a new or refreshed session
@@ -302,13 +336,13 @@ observeEvent(input$get_screen_width, {
 ## check every 5 seconds if width has been reset
 ## https://github.com/rstudio/rstudio/issues/1870
 ## https://community.rstudio.com/t/rstudio-resets-width-option-when-running-shiny-app-in-viewer/3661
-reactivePoll(
-  5000,
-  session,
-  checkFunc = function() {
-    if (getOption("width", default = 250) != 250) options(width = 250)
-  },
-  valueFunc = function() {
-    return()
-  }
-)
+# reactivePoll(
+#   5000,
+#   session,
+#   checkFunc = function() {
+#     if (getOption("width", default = 250) != 250) options(width = 250)
+#   },
+#   valueFunc = function() {
+#     return()
+#   }
+# )